# backend/auth-service/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OAuth 2.0 authorization and user authentication service for the library management system.\nThis service uses opaque bearer tokens, Authorization Code with PKCE, and HttpOnly session cookies.",
        "title": "Library Management Auth API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticates a user and sets an HttpOnly session cookie. HTML form requests are also accepted. A valid returnTo redirects with 303.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in",
                "parameters": [
                    {
                        "description": "Login details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authenticated user",
                        "schema": {
                            "$ref": "#/definitions/response.UserSuccess"
                        }
                    },
                    "303": {
                        "description": "Redirect to authorization flow",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Untrusted request origin",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Invalid login data or return URL",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "description": "Revokes the session identified by the lms_session cookie and clears that cookie.",
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session cookie: lms_session=token",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Session revoked"
                    },
                    "401": {
                        "description": "Invalid session",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Untrusted request origin",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "description": "Returns the user identified by the lms_session cookie.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Get current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session cookie: lms_session=token",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Current user",
                        "schema": {
                            "$ref": "#/definitions/response.UserSuccess"
                        }
                    },
                    "401": {
                        "description": "Invalid session",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Creates a user account. Requests from browsers must use the configured trusted Origin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "Registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Register"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created",
                        "schema": {
                            "$ref": "#/definitions/response.UserSuccess"
                        }
                    },
                    "403": {
                        "description": "Untrusted request origin",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email already registered",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Invalid registration data",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth/userinfo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Get token user",
                "responses": {
                    "200": {
                        "description": "Token user",
                        "schema": {
                            "$ref": "#/definitions/response.UserSuccess"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired access token",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.Login": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 254,
                    "example": "ada@example.com"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "example": "correct horse battery staple"
                },
                "returnTo": {
                    "type": "string",
                    "example": "http://localhost:8081/oauth/authorize"
                }
            }
        },
        "request.Register": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 254,
                    "example": "ada@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Ada Lovelace"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 12,
                    "example": "correct horse battery staple"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "LMS-422001"
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "invalid request data"
                }
            }
        },
...
        "response.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-07-19T08:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "ada@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
                },
                "name": {
                    "type": "string",
                    "example": "Ada Lovelace"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2026-07-19T08:00:00Z"
                }
            }
        },
        "response.UserSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "LMS-200000"
                },
                "data": {
                    "$ref": "#/definitions/response.User"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Prefix access tokens with \"Bearer \".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}